+++ src/interfaces/IJBPriceFeed.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.0;
 
 interface IJBPriceFeed {
-    function currentUnitPrice(uint256 targetDecimals) external view returns (uint256);
+    function currentUnitPrice(uint8 targetDecimals) external view returns (uint160);
 }
