+++ src/interfaces/IJBRulesetApprovalHook.sol
@@ -5,12 +5,12 @@ import {IERC165} from "@openzeppelin/contracts/utils/introspection/IERC165.sol";
 import {JBApprovalStatus} from "./../enums/JBApprovalStatus.sol";
 
 interface IJBRulesetApprovalHook is IERC165 {
-    function DURATION() external view returns (uint256);
+    function DURATION() external view returns (uint40);
 
     function approvalStatusOf(
         uint32 projectId,
-        uint256 rulesetId,
-        uint256 start
+        uint40 rulesetId,
+        uint40 start
     )
         external
         view
